requirements.txt
# Core dependencies
requests>=2.31.0
jinja2>=3.1.0
weasyprint>=61.0
matplotlib>=3.7.0
pyzabbix>=1.3.0

# Optional dependencies for enhanced features
python-dotenv>=1.0.0
pillow>=10.0.0
reportlab>=4.0.0

# LINE Bot integration
line-bot-sdk>=3.5.0

# Development dependencies (optional)
pytest>=7.4.0
black>=23.7.0
flake8>=6.0.0
flask>=2.3.0
flask-cors>=4.0.0
